requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cheroot==8.6.0
click==8.0.4
colorama==0.4.4
dateparser==1.1.0
Flask==2.0.3
Flask-RESTful==0.3.9
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.1
jaraco.functools==3.5.0
Jinja2==3.0.3
MarkupSafe==2.1.0
more-itertools==8.12.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psutil==5.9.0
python-binance==1.0.15
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
unicorn-binance-websocket-api==1.40.2
unicorn-fy==0.12.0
urllib3==1.26.8
websocket-client==1.3.1
websockets==10.2
Werkzeug==2.0.3
yarl==1.7.2
